+++ timeslot_test.go
@@ -82,3 +82,100 @@ func ExampleTimeSlot_3() {
 	// [0] TIME: 2000-01-03 00:00:00 +0100 CET | VALUE: true
 
 }
+
+func ExampleTimeSlot_4() {
+	slot, err := NewTimeSlot("8x7", "00 09 * * *", "00 17 * * *")
+	if err != nil {
+		ilog.Errorf(">>>>USTS ERROR:Error: %s", err)
+		return
+	}
+	loc, err := time.LoadLocation("Europe/Madrid")
+	if err != nil {
+		fmt.Printf("Error: %s", err)
+		return
+	}
+	t0 := time.Date(2000, 1, 1, 17, 0, 0, 0, loc)
+	t1 := time.Date(2000, 1, 1, 18, 0, 0, 0, loc)
+	ilog.Debugf(">>>>USTS DEBUG:%+v", slot)
+	events, err := slot.GetTimeEvents(t0, t1, "Europe/Madrid")
+	if err != nil {
+		fmt.Printf("Error: %s", err)
+	}
+	events.Dump()
+	// Output:
+	// [INIT] Default VALUE: false
+	// [0] TIME: 2000-01-01 17:00:00 +0100 CET | VALUE: false
+}
+
+func ExampleTimeSlot_5() {
+	slot, err := NewTimeSlot("8x7", "00 09 * * *", "00 17 * * *")
+	if err != nil {
+		ilog.Errorf(">>>>USTS ERROR:Error: %s", err)
+		return
+	}
+	loc, err := time.LoadLocation("Europe/Madrid")
+	if err != nil {
+		fmt.Printf("Error: %s", err)
+		return
+	}
+	t0 := time.Date(2000, 1, 1, 8, 0, 0, 0, loc)
+	t1 := time.Date(2000, 1, 1, 18, 0, 0, 0, loc)
+	ilog.Debugf(">>>>USTS DEBUG:%+v", slot)
+	events, err := slot.GetTimeEvents(t0, t1, "Europe/Madrid")
+	if err != nil {
+		fmt.Printf("Error: %s", err)
+	}
+	events.Dump()
+	// Output:
+	//[INIT] Default VALUE: false
+	//[0] TIME: 2000-01-01 09:00:00 +0100 CET | VALUE: true
+	//[1] TIME: 2000-01-01 17:00:00 +0100 CET | VALUE: false
+}
+
+func ExampleTimeSlot_6() {
+	slot, err := NewTimeSlot("8x7", "00 09 * * *", "00 17 * * *")
+	if err != nil {
+		ilog.Errorf(">>>>USTS ERROR:Error: %s", err)
+		return
+	}
+	loc, err := time.LoadLocation("Europe/Madrid")
+	if err != nil {
+		fmt.Printf("Error: %s", err)
+		return
+	}
+	t0 := time.Date(2000, 1, 1, 16, 0, 0, 0, loc)
+	t1 := time.Date(2000, 1, 1, 18, 0, 0, 0, loc)
+	ilog.Debugf(">>>>USTS DEBUG:%+v", slot)
+	events, err := slot.GetTimeEvents(t0, t1, "Europe/Madrid")
+	if err != nil {
+		fmt.Printf("Error: %s", err)
+	}
+	events.Dump()
+	// Output:
+	//[INIT] Default VALUE: true
+	//[0] TIME: 2000-01-01 17:00:00 +0100 CET | VALUE: false
+}
+
+func ExampleTimeSlot_7() {
+	slot, err := NewTimeSlot("8x7", "00 09 * * *", "00 17 * * *")
+	if err != nil {
+		ilog.Errorf(">>>>USTS ERROR:Error: %s", err)
+		return
+	}
+	loc, err := time.LoadLocation("Europe/Madrid")
+	if err != nil {
+		fmt.Printf("Error: %s", err)
+		return
+	}
+	t0 := time.Date(2000, 1, 1, 8, 0, 0, 0, loc)
+	t1 := time.Date(2000, 1, 1, 9, 0, 0, 0, loc)
+	ilog.Debugf(">>>>USTS DEBUG:%+v", slot)
+	events, err := slot.GetTimeEvents(t0, t1, "Europe/Madrid")
+	if err != nil {
+		fmt.Printf("Error: %s", err)
+	}
+	events.Dump()
+	// Output:
+	//[INIT] Default VALUE: false
+	//[0] TIME: 2000-01-01 09:00:00 +0100 CET | VALUE: true
+}