+++ timewindow_test.go
@@ -36,7 +36,7 @@ func ExampleTimeWindow_1() {
 	}
 	ts.DumpInTimezone("Europe/Madrid")
 	// Output:
-	// [INIT] Default VALUE: false
+	// [INIT] Default VALUE: true
 	// [0] TIME: 2020-04-27 00:00:00 +0200 CEST | VALUE: true
 	// [1] TIME: 2020-04-28 00:00:00 +0200 CEST | VALUE: true
 	// [2] TIME: 2020-04-29 00:00:00 +0200 CEST | VALUE: true
@@ -127,7 +127,7 @@ func ExampleTimeWindow_3() {
 	}
 	ts.DumpInTimezone("Europe/Madrid")
 	// Output:
-	// [INIT] Default VALUE: false
+	// [INIT] Default VALUE: true
 	// [0] TIME: 2020-04-27 00:00:00 +0200 CEST | VALUE: true
 	// [1] TIME: 2020-04-28 00:00:00 +0200 CEST | VALUE: true
 	// [2] TIME: 2020-04-29 00:00:00 +0200 CEST | VALUE: true
@@ -183,7 +183,7 @@ func ExampleTimeWindow_4() {
 	}
 	ts.DumpInTimezone("Europe/Madrid")
 	// Output:
-	// [INIT] Default VALUE: false
+	// [INIT] Default VALUE: true
 	// [0] TIME: 2020-04-27 00:00:00 +0200 CEST | VALUE: true
 	// [1] TIME: 2020-04-27 03:00:00 +0200 CEST | VALUE: false
 	// [2] TIME: 2020-04-27 05:00:00 +0200 CEST | VALUE: true
@@ -250,7 +250,7 @@ func ExampleTimeWindow_5() {
 	}
 	ts.DumpInTimezone("Europe/Madrid")
 	// Output:
-	// [INIT] Default VALUE: false
+	// [INIT] Default VALUE: true
 	// [0] TIME: 2020-04-27 00:00:00 +0200 CEST | VALUE: true
 	// [1] TIME: 2020-04-27 03:00:00 +0200 CEST | VALUE: false
 	// [2] TIME: 2020-04-27 05:00:00 +0200 CEST | VALUE: true
@@ -266,4 +266,46 @@ func ExampleTimeWindow_5() {
 	// [12] TIME: 2020-05-04 03:00:00 +0200 CEST | VALUE: false
 	// [13] TIME: 2020-05-04 05:00:00 +0200 CEST | VALUE: true
 	// [14] TIME: 2020-05-05 00:00:00 +0200 CEST | VALUE: true
+}
+
+func ExampleTimeWindow_6() {
+
+	var ts *USTimeSerie
+	var err error
+	var loc *time.Location
+	var slot *TimeSlot
+
+	win := NewTimeWindow("8x5")
+	loc, err = win.SetTimeZone("Europe/Madrid")
+	if err != nil {
+		fmt.Printf("ERROR on set timeZone %s\n", err)
+		return
+	}
+
+	slot, err = NewTimeSlot("8x5_a", "00 09 * * *", "00 13 * * *")
+	if err != nil {
+		fmt.Printf("ERROR on get slot 24x5 %s\n", err)
+		return
+	}
+	win.AddSlot(slot, Add)
+
+	slot, err = NewTimeSlot("8x5_a", "00 14 * * MON-FRI", "00 18 * * MON-FRI")
+	if err != nil {
+		fmt.Printf("ERROR on get slot  mon_3_to_5 %s\n", err)
+		return
+	}
+	win.AddSlot(slot, Add)
+
+	t0 := time.Date(2020, 4, 27, 18, 0, 0, 0, loc)
+	t1 := time.Date(2020, 4, 27, 20, 0, 0, 0, loc)
+
+	ts, err = win.GetTimeEvents(t0, t1)
+	if err != nil {
+		fmt.Printf("ERROR on set timewindow %s\n", err)
+		return
+	}
+	ts.DumpInTimezone("Europe/Madrid")
+	// Output:
+	//[INIT] Default VALUE: false
+	//[0] TIME: 2020-04-27 18:00:00 +0200 CEST | VALUE: false
 }